ref(api): refactor cargos domain

+++ apps/api/internal/cargos/dto.go
@@ -4,10 +4,10 @@ import "time"
 
 type createCargoDTO struct {
 	AddressSequence []string  `json:"addressSequence"`
-	EmployeeID      int       `json:"employeeId"`
-	VehicleID       int       `json:"vehicleId"`
+	EmployeeID      int64     `json:"employeeId"`
+	VehicleID       int64     `json:"vehicleId"`
 	StartDate       time.Time `json:"startDate"`
 	DeadlineDate    time.Time `json:"deadlineDate"`
 	Price           float64   `json:"price"`
-	PaymentStatus   int       `json:"paymentStatus"`
+	PaymentStatus   int64     `json:"paymentStatus"`
 }
